pyproject.toml
[tool.poetry]
name = "madr"
version = "0.1.0"
description = ""
authors = ["Carlos Eduardo <ducameneses16@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
psycopg = {extras = ["binary"], version = "^3.2.1"}
pydantic = {extras = ["email"], version = "^2.8.2"}
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"



[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.1"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.8"
testcontainers = "^4.8.0"
factory-boy = "^3.3.1"
gevent = "^24.2.1"
freezegun = "^1.5.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.coverage.run]
concurrency = ["gevent"]

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_mode = 'auto'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev madr/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=madr -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
